+++ shardingsphere-test/shardingsphere-pipeline-test/src/test/resources/logback-test.xml
@@ -23,7 +23,7 @@
         </encoder>
     </appender>
     <logger name="org.apache.zookeeper" level="off" />
-    <logger name="org.apache.shardingsphere.data.pipeline.scenario.rulealtered.prepare.InventoryTaskSplitter" level="off" />
+    <logger name="org.apache.shardingsphere.data.pipeline.core.prepare.InventoryTaskSplitter" level="off" />
     <logger name="org.apache.shardingsphere.data.pipeline.core.api.impl.RuleAlteredJobAPIImpl" level="off" />
     <logger name="org.apache.shardingsphere.data.pipeline.core.ingest.dumper.AbstractInventoryDumper" level="off" />
